create club

+++ project_api/src/main/java/com/example/project_api/models/repository/ClubRepository.java
@@ -6,6 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface ClubRepository extends JpaRepository<Club, Integer> {
-    Optional<Club> findByTitle(String title);
-    Optional<Club> findByUserId(int userId);
+   Optional<Club> findByTitle(String title);
+   Optional<Club> findByUserId(long userId);
+
 }
